dcp-815 rename terra/terra_listener
