+++ tests/exporter/terra/test_terra_listener.py
@@ -7,7 +7,7 @@ from mock import MagicMock
 from exporter.ingest.service import IngestService
 from exporter.queue.config import QueueConfig
 from exporter.terra.exporter import TerraExporter
-from exporter.terra.terra_listener import _TerraListener
+from exporter.terra.listener import _TerraListener
 
 
 class TerraMessageHandlerTest(TestCase):
